+++ src/pages/Encounters/components/EncounterNoteSkeleton.tsx
@@ -1,0 +1,19 @@
+import { Skeleton } from "@/components/ui/skeleton";
+
+export default function EncounterNoteSkeleton() {
+  return (
+    <div className="space-y-4">
+      {Array.from({ length: 3 }).map((_, i) => (
+        <div key={i} className="p-4 rounded-lg bg-gray-100 animate-pulse">
+          <div className="flex items-start gap-3">
+            <div className="w-8 h-8 rounded-full bg-gray-200" />
+            <div className="flex-1 space-y-2">
+              <Skeleton className="h-4 w-24" />
+              <Skeleton className="h-4 w-3/4" />
+            </div>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+}
