+++ src/pages/FacilityOrganization/components/FacilityOrganizationSkeleton.tsx
@@ -1,0 +1,24 @@
+import { Card, CardFooter, CardHeader } from "@/components/ui/card";
+import { Skeleton } from "@/components/ui/skeleton";
+
+export default function FacilityOrganizationSkeleton() {
+  return (
+    <div className="px-6 py-6 space-y-6">
+      <Skeleton className="h-8 w-48" />
+      <Skeleton className="h-8 w-48 self-end" />
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
+        {Array.from({ length: 6 }).map((_, i) => (
+          <Card key={i} className="relative space-y-4">
+            <CardHeader>
+              <Skeleton className="h-6 w-1/3" />
+              <Skeleton className="h-4 w-1/4" />
+            </CardHeader>
+            <CardFooter>
+              <Skeleton className="h-10 w-full" />
+            </CardFooter>
+          </Card>
+        ))}
+      </div>
+    </div>
+  );
+}
