+++ src/pages/Organization/components/OrganizationLayout.tsx
@@ -15,10 +15,10 @@ import {
 import { Menubar, MenubarMenu, MenubarTrigger } from "@/components/ui/menubar";
 
 import Page from "@/components/Common/Page";
-import { OrganizationIndexSkeleton } from "@/components/Common/SkeletonComponents";
 
 import query from "@/Utils/request/query";
 import { usePermissions } from "@/context/PermissionContext";
+import OrganizationLayoutSkeleton from "@/pages/Organization/components/OrganizationLayoutSkeleton";
 import {
   Organization,
   OrganizationParent,
@@ -69,7 +69,7 @@ export default function OrganizationLayout({
   }, [org, setOrganization]);
 
   if (isLoading) {
-    return <OrganizationIndexSkeleton />;
+    return <OrganizationLayoutSkeleton />;
   }
   // add loading state
   if (!org) {
